Enhance login UI and ViewModel linkage
This commit introduces several key changes aimed at improving the user experience and functionality of the login interface. In the AccountViewModel.cs, the AccountName property of the InputWindow's ViewModel is now set to the UserName, ensuring that the account name is pre-populated or utilized in the input window. This change enhances user convenience by reducing the need for manual input.

Furthermore, an OkAction has been defined in the InputWindow's ViewModel to create a SecureString from the input text, enhancing security measures without altering existing behavior.

In InputViewModel.cs, the default state of the rememberPassword field has been changed to true, making the "Remember Password" option enabled by default. This adjustment reflects a more user-friendly approach to password management. Additionally, a new AccountName property and a dependent Description property have been introduced. The Description property dynamically generates a message that prompts the user to enter the password for the specified account, improving the clarity and relevance of UI prompts.

InputWindow.xaml has undergone several modifications to support these backend changes. The window's minimum height is set to 200, and it is centered on the screen upon startup, improving the initial appearance and positioning. The window title has been updated from "Credentials" to "Login", signaling a clearer purpose of the window. The UI now binds the Label content to the Description property, allowing for dynamic display of account-specific messages. The text on the checkbox has been revised to "Remember the password", clarifying its function. Lastly, adjustments to the margins of various UI elements have been made to refine the overall layout and appearance.

These changes collectively enhance the login process, making it more intuitive, secure, and user-friendly.